pwr/power-market-analytics/requirements.txt
django>=4.2
djangorestframework>=3.14
pandas>=2.0
numpy>=1.24
scikit-learn>=1.2
joblib>=1.2
sqlalchemy>=2.0
python-dotenv>=1.0
PyMySQL>=1.0